% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\docType{class}
\name{BetaUniformModel-class}
\alias{BetaUniformModel-class}
\title{Object representation of a beta-uniform model fit to a P-value distribution}
\description{
Details the model P ~ (1-λ)β(a,1) + λβ(1,1), fit to a P-value set
}
\section{Slots}{

\describe{
\item{\code{a}}{Shape parameter of distribution β(a,1)}

\item{\code{lambda}}{Mixture parameter}

\item{\code{pvalues}}{A numeric vector of the values fit by the model}

\item{\code{negLLH}}{The negative log-likelihood of the model}
}}

