% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaUniformPvalueSumTest.R
\name{betaUniformPvalueSumTest}
\alias{betaUniformPvalueSumTest}
\alias{betaUniformPvalueSumTest,numeric,BetaUniformModel-method}
\alias{betaUniformPvalueSumTest,Pvalues,BetaUniformModel-method}
\title{A meta-analysis P-value summarising a set of hypotheses}
\usage{
betaUniformPvalueSumTest(testPvalues, betaUniformFit, na.rm = TRUE)

\S4method{betaUniformPvalueSumTest}{numeric,BetaUniformModel}(testPvalues, betaUniformFit, na.rm = TRUE)

\S4method{betaUniformPvalueSumTest}{Pvalues,BetaUniformModel}(testPvalues, betaUniformFit, na.rm = TRUE)
}
\arguments{
\item{testPvalues}{A numerical vector of P-values to be tested.}

\item{betaUniformFit}{A beta-uniform model of the P-value distribution.}

\item{...}{Unused. Included to allow for similar form across P-value sum tests.}
}
\value{
combinedPval A single p-value combining the results of multiple hypothesis tests into one
}
\description{
A variant of Fisher's method for P-value aggregation, but with a parametrically derived null-hypothesis that reflects the background signal in the data
}
\details{
Authors in the literature (e.g. Luo et al.) have used a chi-squared test for gene set enrichment with the null-hypothesis that all genes
have P-values derived from a pure noise (uniform) distribution. This is typically demonstrably not true and will result in an inflated rate
of false positives
}
\section{Methods (by class)}{
\itemize{
\item \code{testPvalues = numeric,betaUniformFit = BetaUniformModel}: Convert to P-values on the fly

\item \code{testPvalues = Pvalues,betaUniformFit = BetaUniformModel}: Compute a P-value with the null-hypothesis being that they are drawn randomly from a mixture distribution
}}

\references{
Luo, W., Friedman, M. S., Shedden, K., Hankenson, K. D., & Woolf, P. J. (2009). GAGE: generally applicable gene set enrichment for pathway analysis. BMC Bioinformatics

\url{https://en.wikipedia.org/wiki/Phase-type_distribution}

Sidje RB. Expokit: a software package for computing matrix exponentials. ACM Trans Math Softw. 1998;24: 130–156.
}
\seealso{
fishersCombindedProbabilityTest

fitBetaUniformParameters

BioNet::fitBumModel

Other P-value sum tests: 
\code{\link{fishersPvalueSumTest}()}
}
\concept{P-value sum tests}
