% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaUniformScore.R
\name{betaUniformScore}
\alias{betaUniformScore}
\alias{betaUniformScore,Pvalues,BetaUniformModel,ScalarNumeric-method}
\alias{betaUniformScore,igraph,ANY,ANY-method}
\alias{betaUniformScore,numeric,ANY,ANY-method}
\alias{betaUniformScore,ANY,ANY,numeric-method}
\alias{betaUniformScore,BetaUniformModel,missing,ANY-method}
\alias{betaUniformScore,Pvalues,NULL,ANY-method}
\alias{betaUniformScore,ANY,bum,ANY-method}
\title{For each P-value, compute the log likelihood ratio of originating from signal or noise component}
\usage{
betaUniformScore(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{Pvalues,BetaUniformModel,ScalarNumeric}(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{igraph,ANY,ANY}(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{numeric,ANY,ANY}(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{ANY,ANY,numeric}(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{BetaUniformModel,missing,ANY}(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{Pvalues,`NULL`,ANY}(x, betaUniformFit = NULL, FDR = 0.05)

\S4method{betaUniformScore}{ANY,bum,ANY}(x, betaUniformFit = NULL, FDR = 0.05)
}
\arguments{
\item{x}{data}

\item{betaUniformFit}{(optional) A beta-uniform model of the P-value distribution. Set to NULL to autofit on the fly.}

\item{FDR}{The tolerable fraction of false positives within the set of positive scoring values. See Morris & Pounds (2003)}
}
\value{
betaUniformScores A vector of P-Value scores
}
\description{
Relative to some false discovery threshold, compute the relative
ratio of probabilities and then take the natural logarithm. This function is
somewhat analagous to BioNet::scoreNodes
}
\section{Functions}{
\itemize{
\item \code{betaUniformScore,Pvalues,BetaUniformModel,ScalarNumeric-method}: Score P-values against an explicit beta-uniform model object

\item \code{betaUniformScore,igraph,ANY,ANY-method}: Score nodes in an igraph; there must be a 'pValue'-like attribute (pVal, p.value etc.)

\item \code{betaUniformScore,numeric,ANY,ANY-method}: Score nodes from numeric object

\item \code{betaUniformScore,ANY,ANY,numeric-method}: Score nodes default

\item \code{betaUniformScore,BetaUniformModel,missing,ANY-method}: Score p-values from a BetaUniformModel object against inself

\item \code{betaUniformScore,Pvalues,NULL,ANY-method}: Score P-values against a Beta-uniform model fit on the fly.

\item \code{betaUniformScore,ANY,bum,ANY-method}: Score P-values against a beta-uniform model S3 object from BioNet
}}

\references{
Pounds, S., & Morris, S. W. (2003). Estimating the occurrence of false positives and false negatives in microarray studies by approximating and partitioning the empirical distribution of p-values. Bioinformatics

Dittrich MT, Klau GW, Rosenwald A, Dandekar T, Müller T. Identifying functional modules in protein-protein interaction networks: An integrated exact approach. Bioinformatics. 2008
}
\seealso{
fitBetaUniformParameters BioNet::scoreNodes
}
