% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioplanetPathwaysDT.R
\docType{data}
\name{bioplanetPathwaysDT}
\alias{bioplanetPathwaysDT}
\title{Bioplanet - catalogue of biological pathways}
\format{
A data frame with 74148 rows and 5 variables:
\describe{
\item{BioPlanetName}{BioPlanetName, names of the pathway}
\item{BioPlanetID}{"BioPlanetID, ID of the pathway}
\item{geneID}{geneID, gene ID}
\item{geneSymbol}{"geneSymbol, gene symbol}
\item{category}{category, category}
}
}
\source{
\url{https://tripod.nih.gov/bioplanet/}
}
\usage{
bioplanetPathwaysDT
}
\description{
A dataset includind biological pathways from Bioplanet
}
\keyword{datasets}
