% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaUniformPvalueSumTest.R
\name{constructHyperexponentialSumTransitionMatrix}
\alias{constructHyperexponentialSumTransitionMatrix}
\title{Model a sum of log-transformed beta-uniform mixture draws as a sum of hyperexponential distributions: a phase-type distribution}
\usage{
constructHyperexponentialSumTransitionMatrix(
  nValues,
  uniformProportion,
  fittedBetaShape
)
}
\description{
The distribution of a log-trasnformed beta-uniform draw is a hyperexponential distribution. The distribution of the sum of multiple
such hyperexponentials does nto have a name, but can be treated as a phase-type distribution with appropriate transition matrix.
}
\references{
\url{https://en.wikipedia.org/wiki/Phase-type_distribution}
}
