% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{convertFromBUM}
\alias{convertFromBUM}
\title{A simple adapter from the class produced by BioNet}
\usage{
convertFromBUM(obj)
}
\description{
A simple adapter from the class produced by BioNet
}
