% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBetaUniformModel.R
\name{dbetauniform}
\alias{dbetauniform}
\title{Density function for Beta-uniform Model}
\usage{
dbetauniform(x, a, lambda, log = FALSE)
}
\arguments{
\item{x}{vector of values}

\item{a}{beta distribution shape parameter}

\item{lambda}{parameter lambda}

\item{log}{return the logarithmic version of the output}
}
\description{
Density function for Beta-uniform Model
}
\references{
Pounds, S., & Morris, S. W. (2003). Estimating the occurrence of false positives and false negatives in microarray studies by approximating and partitioning the empirical distribution of p-values. Bioinformatics
}
