% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erlotinib_adrHCC827_diffexDT.R
\docType{data}
\name{erlotinib_adrHCC827_diffexDT}
\alias{erlotinib_adrHCC827_diffexDT}
\title{Microarray assay to understand acquired drug resistance to EGFR-targeted therapy (erlotinib) in cell lines}
\format{
A data.table with 48,803 rows, documenting probe-level differential expression in three erlotinib-treated cell lines: HCC827 (susceptible) and against ER3/T15.2 (evolved resistance). Additional gene symbols, entrez gene IDs and uniprot annotations also included.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE38310}
}
\usage{
data(erlotinib_adrHCC827_diffexDT)
}
\description{
GSE38310: HCC827 cell lines, which are susceptible to erlotinib, were treated to induce resistence in vivo by repeated exposure. Gene expression of cell lines
treated with erlotibin was collected (alongside controls) in both susceptible and resistant cell lines. Expression captured using Illumina HumanHT-12 V3.0 expression beadchip.
}
\details{
The study authors cocluded that AXL kinase caused resistance to EGFR therapy. A clean peturbagenic experiment.

Generated 10 December 2019 from raw expression values dowloaded from GEO. Values were normalised using limma::neqc and processed with further limma functions.
}
\references{
\url{https://en.wikipedia.org/wiki/Erlotinib}

Zhang Z, Lee JC, Lin L, Olivas V et al. Activation of the AXL kinase causes resistance to EGFR-targeted therapy in lung cancer. Nat Genet 2012
}
\keyword{datasets}
