% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{falseNegativeFraction}
\alias{falseNegativeFraction}
\title{False negative rate from M&P: B table 1 and equation 5}
\usage{
falseNegativeFraction(obj, pValueThreshold)
}
\arguments{
\item{obj}{data object}

\item{pValueThreshold}{p-value threshold}
}
\description{
False negative rate from M&P: B table 1 and equation 5
}
