% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{falsePositiveFraction}
\alias{falsePositiveFraction}
\title{FALSE positive rate from M&P: C table 1 and equation 5}
\usage{
falsePositiveFraction(obj, pValueThreshold)
}
\arguments{
\item{obj}{data object}

\item{pValueThreshold}{p-value threshold}
}
\description{
FALSE positive rate from M&P: C table 1 and equation 5
}
