% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishersPvalueSumTest.R
\name{fishersPvalueSumTest}
\alias{fishersPvalueSumTest}
\alias{fishersPvalueSumTest,numeric-method}
\alias{fishersPvalueSumTest,Pvalues-method}
\title{Combine multiple independent P-values together using Fisher's method}
\usage{
fishersPvalueSumTest(testPvalues, ...)

\S4method{fishersPvalueSumTest}{numeric}(testPvalues, ...)

\S4method{fishersPvalueSumTest}{Pvalues}(testPvalues, na.rm = TRUE, rate = 1, ...)
}
\arguments{
\item{testPvalues}{A numerical vector of P-values to be tested.}

\item{...}{Unused. Included to allow for similar form across P-value sum tests.}

\item{rate}{Provide the rate for the expoential in the sum (distributed as a gamma) \link{optional}}
}
\value{
combinedPval A single P-value combining the results of multiple independent hypothesis tests into one
}
\description{
Fisher's method for combining P-values together consists of studying a test statistic of \eqn{T = \sum_i -log(P_i)} and
finding the corresponding point in the cumulative distribution of an appropriately parameterised chi-squared distribution.

The basic assumption is that P-values are uniformly distributed under the null-hypothesis, meaning that \eqn{-log(P) \sim exp(1)}. Recall that
if \eqn{X \sim exp(1)} then \eqn{\sum_{k} X_{k} \sim \Gamma (k,1) \sim \chi^2 (2k) } (note that the gamma and erlang distributions are identical for integer \eqn{k}).
We therefore simply calculate the point in the cumulative of the gamma distribution
}
\details{
Note that the use of this method for analysis of omics datasets was pioneered by Luo et al.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Convert to P-values on the fly

\item \code{Pvalues}: Fisher's combined probability test
}}

\references{
\url{https://en.wikipedia.org/wiki/Fisher\%27s_method}

\url{https://en.wikipedia.org/wiki/Erlang_distribution}

Luo, W., Friedman, M. S., Shedden, K., Hankenson, K. D., & Woolf, P. J. (2009). GAGE: generally applicable gene set enrichment for pathway analysis. BMC Bioinformatics
}
\seealso{
betaUniformPvalueSumTest

Other P-value sum tests: 
\code{\link{betaUniformPvalueSumTest}()}
}
\concept{P-value sum tests}
