% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsPathBankProteinSets_DT.R
\docType{data}
\name{hsPathBankProteinSets_DT}
\alias{hsPathBankProteinSets_DT}
\title{Pathbank - catalogue of biological pathways}
\format{
A data frame with 303334 rows and 7 variables:
\describe{
\item{pathbank}{pathbank, pathway ID}
\item{pathway}{"pathway, name of the pathway}
\item{pathwayType}{pathwayType, pathway type}
\item{accession}{"accession, protein accession}
\item{geneSymbol}{geneSymbol, gene symbol}
\item{geneID}{geneID, gene ID}
\item{ENSG}{ENSG, ensembl gene ID}
}
}
\source{
\url{https://www.pathbank.org/}
}
\usage{
hsPathBankProteinSets_DT
}
\description{
Pathbank is a new (Oct 2019) resource from the people that made SMPD
(The Small Molecule Pathway Database).
}
\keyword{datasets}
