% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDEG.R
\name{metaDEG}
\alias{metaDEG}
\alias{metaDEG,data.frame,data.frame,BetaUniformModel-method}
\alias{metaDEG,data.table,data.table,missing-method}
\alias{metaDEG,ANY,list,ANY-method}
\alias{metaDEG,numeric,ANY,ANY-method}
\title{meta differential gene expression (DEG) analysis}
\usage{
metaDEG(pValueSet, geneSet, betaUniformFit, ...)

\S4method{metaDEG}{data.frame,data.frame,BetaUniformModel}(
  pValueSet,
  geneSet,
  betaUniformFit,
  pValAttr = "pVal",
  geneSetNameAttr = "geneSet",
  geneSetMembersAttr = "gene"
)

\S4method{metaDEG}{data.table,data.table,missing}(pValueSet, geneSet, betaUniformFit, pValAttr, plot = TRUE, ...)

\S4method{metaDEG}{ANY,list,ANY}(
  pValueSet,
  geneSet,
  geneSetNameAttr = "geneSet",
  geneSetMembersAttr = "gene",
  ...
)

\S4method{metaDEG}{numeric,ANY,ANY}(
  pValueSet,
  geneSet,
  pValAttr = "pVal",
  geneSetMembersAttr = "gene",
  ...
)
}
\arguments{
\item{pValueSet}{P-values from experimental assay}

\item{geneSet}{Gene set collections}

\item{betaUniformFit}{A beta-uniform model of the P-value distribution -
optional - can be estimated from the P-Values provided}

\item{...}{Additional parameters for methods}

\item{pValAttr}{The name of the P-value attribute in the pValueSet frame}

\item{geneSetNameAttr}{The name of the gene set identifier attribute in the geneSet frame}

\item{geneSetMembersAttr}{The name of the gene set members attribute in the geneSet frame}

\item{plot}{Logical - draw a plot or not - default TRUE}
}
\value{
A frame of gene sets, their members, their sizes and their meta-analysis P-values
}
\description{
Given a collection of gene sets and
}
\section{Functions}{
\itemize{
\item \code{metaDEG,data.frame,data.frame,BetaUniformModel-method}: Build a beta-uniform model on the fly

\item \code{metaDEG,data.table,data.table,missing-method}: Build a betaUniform model on the fly

\item \code{metaDEG,ANY,list,ANY-method}: Provide geneSet as a named list of members

\item \code{metaDEG,numeric,ANY,ANY-method}: Provide pValueSet as a uniquely named numeric vector of pvalues
}}

