% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDEGth-package.R
\docType{package}
\name{metaDEGth}
\alias{metaDEGth}
\title{metaDEGth: Combining differential expression P-values in a statistically consistent fashion whilst respecting background signal/noise trends}
\author{
Adam Sardar
}
