% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{noiseFractionUpperBound}
\alias{noiseFractionUpperBound}
\title{Upper bound on the the noise component of mixture distribution}
\usage{
noiseFractionUpperBound(obj)
}
\arguments{
\item{obj}{data object}
}
\description{
π in equation 3 from Morris & Pounds
}
