% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BetaUniformModel.R
\name{plot.BetaUniformModel}
\alias{plot.BetaUniformModel}
\title{Plot summary for fitted Beta-Uniform models}
\usage{
\method{plot}{BetaUniformModel}(
  x,
  showFit = TRUE,
  outputParameters = TRUE,
  outputFormula = TRUE,
  studyName = NULL,
  ...
)
}
\arguments{
\item{x}{A BetaUniformModel fit of data}

\item{showFit}{Flag to specify if you would like the Beta-Uniform model formula outputted as well (default: TRUE)}

\item{outputParameters}{Flag to specify if the plot should list the fitted parameters for the model (defualt:TRUE)}

\item{outputFormula}{Flag to specify if you would like the Beta-Uniform model formula outputted as well (default: TRUE)}

\item{studyName}{Subtitle for plots (default: variable name of x passed in)}

\item{...}{Ignored}
}
\value{
A named list with three ggplots: p-value distribution; QQplot of P-values against a the fitted mixture model and liklihood surface of paramter estimates.
}
\description{
Produce a list of summary ggplots. The default print method produces a grid panel, or you can extract and use them at your wish.
}
\details{
These should assist in diagnosing good or poor model fit.
}
