% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBetaUniformModel.R
\name{qbetauniform}
\alias{qbetauniform}
\title{Quantile function for Beta-uniform model}
\usage{
qbetauniform(p, a, lambda)
}
\arguments{
\item{p}{vector of probabilities}

\item{a}{non-negative parameter of the Beta distribution.}

\item{lambda}{lambda parameter}
}
\description{
Quantile function for Beta-uniform model
}
\references{
Pounds, S., & Morris, S. W. (2003). Estimating the occurrence of false positives and false negatives in microarray studies by approximating and partitioning the empirical distribution of p-values. Bioinformatics
}
