% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{show,BetaUniformModel-method}
\alias{show,BetaUniformModel-method}
\title{Simple summary printer for a beta-uniform fit}
\usage{
\S4method{show}{BetaUniformModel}(object)
}
\description{
Simple summary printer for a beta-uniform fit
}
