% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{trueNegativeFraction}
\alias{trueNegativeFraction}
\title{True negative rate from M&P: D table 1 and equation 5}
\usage{
trueNegativeFraction(obj, pValueThreshold)
}
\description{
True negative rate from M&P: D table 1 and equation 5
}
