% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTRUST_TFtarget_relationships.R
\docType{data}
\name{TTRUST_TF2targets_DT}
\alias{TTRUST_TF2targets_DT}
\title{A Manually Curated Dataset Of Human Transcription Factor - Target Relationships}
\format{
A data.table detailing 795 transcription factors and 2,492 targets. Some target entries are duplicated on account of many-to-one mappings to ensembl identifiers.
}
\source{
\url{https://www.grnpedia.org/trrust/}
}
\usage{
data(TTRUST_TF2targets_DT)
}
\description{
The TRRUST (Transcriptional Regulatory Relationships Unraveled by Sentence-based Text mining) resource is a semi-automated database of 
transcription factor relationships. It has been augmented with ensembl gene identifiers.
}
\details{
TTRUST is distributed under a Creative Commons Attribution-ShareAlike license. You are free to distribute this data independently of the package license, in keeping with that of creative commons.
}
\references{
\url{https://creativecommons.org/licenses/by-sa/4.0/}

Han H, Cho J-W, Lee S, Yun A, Kim H, Bae D, et al. TRRUST v2: an expanded reference database of human and mouse transcriptional regulatory interactions. Nucleic Acids Res. 2018
}
\keyword{datasets}
