% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBetaUniformModel.R
\name{uniformComponentBound}
\alias{uniformComponentBound}
\title{Upper bound on Uniform component of Beta-uniform model}
\usage{
uniformComponentBound(x, a, lambda)
}
\description{
Upper bound on Uniform component of Beta-uniform model
}
\references{
Pounds, S., & Morris, S. W. (2003). Estimating the occurrence of false positives and false negatives in microarray studies by approximating and partitioning the empirical distribution of p-values. Bioinformatics
}
