% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{FDRestimate}
\alias{FDRestimate}
\title{False discovery rate - of BH fame. Equation 6 from M&P}
\usage{
FDRestimate(obj, pValueThreshold)
}
\arguments{
\item{obj}{data object}

\item{pValueThreshold}{p-value threshold}
}
\description{
False discovery rate - of BH fame. Equation 6 from M&P
}
