% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{pValueThresholdAtConfidence}
\alias{pValueThresholdAtConfidence}
\title{pValueThreshold estimate - estimates a threshdold such that FDRestimate(obj, pValueThreshold) <= confidenceLevel. Equatuion 7 in Morris & Pounds}
\usage{
pValueThresholdAtConfidence(obj, confidenceLevel)
}
\arguments{
\item{obj}{data object}

\item{confidenceLevel}{confidence level}
}
\description{
pValueThreshold estimate - estimates a threshdold such that FDRestimate(obj, pValueThreshold) <= confidenceLevel. Equatuion 7 in Morris & Pounds
}
