% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBetaUniformModel.R
\name{rbetauniform}
\alias{rbetauniform}
\title{Random draw function for Beta-uniform model}
\usage{
rbetauniform(n, a, lambda)
}
\description{
Random draw function for Beta-uniform model
}
\references{
Pounds, S., & Morris, S. W. (2003). Estimating the occurrence of false positives and false negatives in microarray studies by approximating and partitioning the empirical distribution of p-values. Bioinformatics
}
