% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaUniformModel_S4Class.R
\name{truePositiveFraction}
\alias{truePositiveFraction}
\title{True positive rate from M&P: A in table 1 and equation 5}
\usage{
truePositiveFraction(obj, pValueThreshold)
}
\description{
True positive rate from M&P: A in table 1 and equation 5
}
